Initial Commit
These files are all that is needed to run the Final_GUI_27 (named as
such because I had to change the python version I was working with to
python 2.7) The actual algorithm and user interface will take about 15
seconds to load, while the algorithm actually generates the schedule.
This is a very complex scheduler, accounting for over 30 different
parameters, for 20 different groups and 12  camp periods over the course
of 1 week. It also factors in the frequency of an activity for a group
in relation to that activity for all other groups, which also adds into
the complexity.
This program was used to generate schedules, starting in the summer of
2013 and saved between 3 and 4 hours worth of work for the person who
had to create the schedules. Hopefully, further versions will continue
to increase the amount of time saved, as well as the efficiency of the
program.